SerpAPI/requirements.txt
simple-salesforce==1.12.5
google-search-results==2.4.2
python-dotenv==1.0.0
pandas==2.2.0
requests==2.31.0
fuzzywuzzy
python-Levenshtein
supabase

